+++ src/cmakeTaskProvider.ts
@@ -17,7 +17,7 @@ import * as util from '@cmt/util';
 import * as expand from '@cmt/expand';
 import { CommandResult } from 'vscode-cmake-tools';
 import { CompileOutputConsumer } from '@cmt/diagnostics/build';
-import { BuildColorMode, colorizeBuildLine } from '@cmt/colorize';
+import { BuildColorMode, GlyphStyle, decorateBuildLine } from '@cmt/colorize';
 import collections from '@cmt/diagnostics/collections';
 import { addDiagnosticToCollection, diagnosticSeverity, populateCollection } from '@cmt/diagnostics/util';
 
@@ -372,6 +372,7 @@ export class CustomBuildTaskTerminal extends proc.CommandConsumer implements vsc
     // How build-tool output is colorized in this terminal. Only set for build
     // tasks (see runBuildTask); stays 'off' for config/test/package/workflow.
     private colorMode: BuildColorMode = 'off';
+    private glyphStyle: GlyphStyle = 'unicode';
     public get onDidWrite(): vscode.Event<string> {
         return this.writeEmitter.event;
     }
@@ -386,12 +387,12 @@ export class CustomBuildTaskTerminal extends proc.CommandConsumer implements vsc
     // These two override methods are used to write output and error messages to the terminal, as well
     // as call the parent class's output and error methods, which store the stdout and stderr messages for returning that info later.
     override output(line: string): void {
-        this.writeEmitter.fire(colorizeBuildLine(line, this.colorMode) + endOfLine);
+        this.writeEmitter.fire(decorateBuildLine(line, this.colorMode, this.glyphStyle) + endOfLine);
         super.output(line);
     }
 
     override error(error: string): void {
-        this.writeEmitter.fire(colorizeBuildLine(error, this.colorMode) + endOfLine);
+        this.writeEmitter.fire(decorateBuildLine(error, this.colorMode, this.glyphStyle) + endOfLine);
         super.error(error);
     }
 
@@ -564,6 +565,7 @@ export class CustomBuildTaskTerminal extends proc.CommandConsumer implements vsc
         if (cmakeDriver) {
             cmakePath = cmakeDriver.getCMakeCommand();
             this.colorMode = cmakeDriver.config.colorizedBuildOutput;
+            this.glyphStyle = cmakeDriver.config.buildOutputGlyphs;
 
             if (!this.options) {
                 this.options = {};