+++ src/config.ts
@@ -193,7 +193,8 @@ export interface ExtensionConfigurationSettings {
     saveBeforeBuild: boolean;
     buildBeforeRun: boolean;
     clearOutputBeforeBuild: boolean;
-    colorizedBuildOutput: "off" | "severity";
+    colorizedBuildOutput: "off" | "severity" | "rich";
+    buildOutputGlyphs: "unicode" | "ascii";
     configureSettings: { [key: string]: boolean | number | string | string[] | util.CMakeValue };
     cacheInit: string | string[] | null;
     preferredGenerators: string[];
@@ -393,8 +394,11 @@ export class ConfigurationReader implements vscode.Disposable {
     get clearOutputBeforeBuild(): boolean {
         return !!this.configData.clearOutputBeforeBuild;
     }
-    get colorizedBuildOutput(): "off" | "severity" {
+    get colorizedBuildOutput(): "off" | "severity" | "rich" {
         return this.configData.colorizedBuildOutput;
+    }
+    get buildOutputGlyphs(): "unicode" | "ascii" {
+        return this.configData.buildOutputGlyphs;
     }
     get configureSettings(): {[key: string]: boolean | number | string | string[] | util.CMakeValue} {
         return this.configData.configureSettings;
@@ -716,7 +720,8 @@ export class ConfigurationReader implements vscode.Disposable {
         saveBeforeBuild: new vscode.EventEmitter<boolean>(),
         buildBeforeRun: new vscode.EventEmitter<boolean>(),
         clearOutputBeforeBuild: new vscode.EventEmitter<boolean>(),
-        colorizedBuildOutput: new vscode.EventEmitter<"off" | "severity">(),
+        colorizedBuildOutput: new vscode.EventEmitter<"off" | "severity" | "rich">(),
+        buildOutputGlyphs: new vscode.EventEmitter<"unicode" | "ascii">(),
         configureSettings: new vscode.EventEmitter<{ [key: string]: any }>(),
         cacheInit: new vscode.EventEmitter<string | string[] | null>(),
         preferredGenerators: new vscode.EventEmitter<string[]>(),