+++ test/unit-tests/config.test.ts
@@ -14,6 +14,7 @@ function createConfig(conf: Partial<ExtensionConfigurationSettings>): Configurat
         buildBeforeRun: true,
         clearOutputBeforeBuild: true,
         colorizedBuildOutput: "off",
+        buildOutputGlyphs: "unicode",
         configureSettings: {},
         cacheInit: null,
         preferredGenerators: [],